requirements.txt
scapy==2.5.0
netifaces==0.11.0
python-dotenv==1.0.0
colorama==0.4.6 